# lib/api-zod/src/generated/types/createYardInspectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateYardInspectionBodyType } from "./createYardInspectionBodyType";

export interface CreateYardInspectionBody {
  vehicleId: number;
  type: CreateYardInspectionBodyType;
  locationId?: number;
  notes?: string;
  bodyDamage?: string;
  fuelPercentage?: number;
}