# lib/api-zod/src/generated/types/createYardLocationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateYardLocationBodyType } from "./createYardLocationBodyType";

export interface CreateYardLocationBody {
  name: string;
  type: CreateYardLocationBodyType;
  city: string;
  address?: string;
  totalCapacity: number;
}