# lib/api-zod/src/generated/types/yardInspection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { YardInspectionStatus } from "./yardInspectionStatus";
import type { YardInspectionType } from "./yardInspectionType";

export interface YardInspection {
  id: number;
  inspectionNumber: string;
  vehicleId: number;
  stockVin: string;
  vehicleName: string;
  type: YardInspectionType;
  status: YardInspectionStatus;
  locationId?: number | null;
  locationName?: string | null;
  notes?: string | null;
  bodyDamage?: string | null;
  fuelPercentage?: number | null;
  createdAt: string;
  completedAt?: string | null;
}