# lib/api-zod/src/generated/types/yardLocationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { YardLocation } from "./yardLocation";
import type { YardZone } from "./yardZone";

export type YardLocationDetail = YardLocation & {
  zones: YardZone[];
};