# lib/api-zod/src/generated/types/yardVehicle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { YardVehicleCondition } from "./yardVehicleCondition";
import type { YardVehicleStatus } from "./yardVehicleStatus";

export interface YardVehicle {
  id: number;
  vin: string;
  stockNumber: string;
  make: string;
  model: string;
  year: number;
  color?: string | null;
  mileage?: number | null;
  condition?: YardVehicleCondition;
  status: YardVehicleStatus;
  locationId?: number | null;
  locationName?: string | null;
  spotId?: number | null;
  spotCode?: string | null;
  zoneName?: string | null;
  price?: number | null;
  imageUrl?: string | null;
  arrivedAt?: string | null;
}